Update resume/hugo.yaml configuration
This commit includes a significant overhaul of the resume/hugo.yaml file. Unnecessary comments, whitespace, and some disabled feature configurations are removed to enhance readability. The navigation menu headers are commented out, the theme toggle is disabled, and search functionality is deactivated. Updates are also made to the `navbar` configuration, disabling hero and other sections, while adding and configuring the `experience` section. Some contact details are edited as well.