Update blog post and enable blog menu link
Renamed the existing 'placeholder-text.md' to 'hello-world.md' with updated content in the blog post. The blog post is titled 'Getting Started with GoHugo: A Developer's Guide'. Enabled the 'Blog' link in the main menu of 'hugo.yaml', updating the URL to '/blog'. Also enabled the 'Recent Posts' section in the footer.